refactor: `constant` 패키지 추가, 검증 패턴 및 메세지 문자열 상수 class로 분리
