+++ backend/src/main/java/com/woowacourse/moragora/constant/ValidationMessages.java
@@ -1,0 +1,7 @@
+package com.woowacourse.moragora.constant;
+
+public class ValidationMessages {
+
+    public static final String MISSING_REQUIRED_INPUT = "필수 입력 값이 누락됐습니다.";
+    public static final String INVALID_FORMAT = "입력 형식이 올바르지 않습니다.";
+}
