+++ backend/src/main/java/com/woowacourse/moragora/dto/request/auth/EmailRequest.java
@@ -1,5 +1,7 @@
 package com.woowacourse.moragora.dto.request.auth;
 
+import com.woowacourse.moragora.constant.Patterns;
+import com.woowacourse.moragora.constant.ValidationMessages;
 import javax.validation.constraints.Email;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -10,7 +12,7 @@ import lombok.ToString;
 @ToString
 public class EmailRequest {
 
-    @Email(regexp = "^[a-zA-Z0-9+-\\_.]+@[a-zA-Z0-9-]+\\.[a-zA-Z0-9-.]+$", message = "입력 형식이 올바르지 않습니다.")
+    @Email(regexp = Patterns.EMAIL, message = ValidationMessages.INVALID_FORMAT)
     private String email;
 
     public EmailRequest(final String email) {
