+++ backend/src/main/java/com/woowacourse/moragora/dto/request/event/EventCancelRequest.java
@@ -1,5 +1,7 @@
 package com.woowacourse.moragora.dto.request.event;
 
+import com.woowacourse.moragora.constant.Patterns;
+import com.woowacourse.moragora.constant.ValidationMessages;
 import java.time.LocalDate;
 import java.util.List;
 import javax.validation.constraints.NotNull;
@@ -13,10 +15,8 @@ import org.springframework.format.annotation.DateTimeFormat;
 @ToString
 public class EventCancelRequest {
 
-    private static final String MISSING_REQUIRED_INPUT = "필수 입력 값이 누락됐습니다.";
-
-    @NotNull(message = MISSING_REQUIRED_INPUT)
-    @DateTimeFormat(pattern = "yyyy-MM-dd")
+    @NotNull(message = ValidationMessages.MISSING_REQUIRED_INPUT)
+    @DateTimeFormat(pattern = Patterns.DATE)
     private List<LocalDate> dates;
 
     public EventCancelRequest(final List<LocalDate> dates) {
