+++ backend/src/main/java/com/woowacourse/moragora/dto/request/meeting/MasterRequest.java
@@ -1,5 +1,6 @@
 package com.woowacourse.moragora.dto.request.meeting;
 
+import com.woowacourse.moragora.constant.ValidationMessages;
 import javax.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -9,9 +10,8 @@ import lombok.ToString;
 @Getter
 @ToString
 public class MasterRequest {
-    private static final String MISSING_REQUIRED_INPUT = "필수 입력 값이 누락됐습니다.";
 
-    @NotNull(message = MISSING_REQUIRED_INPUT)
+    @NotNull(message = ValidationMessages.MISSING_REQUIRED_INPUT)
     private Long userId;
 
     public MasterRequest(final Long userId) {
