+++ backend/src/main/java/com/woowacourse/moragora/dto/request/meeting/MeetingUpdateRequest.java
@@ -1,5 +1,6 @@
 package com.woowacourse.moragora.dto.request.meeting;
 
+import com.woowacourse.moragora.constant.ValidationMessages;
 import javax.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -12,9 +13,8 @@ import org.hibernate.validator.constraints.Length;
 public class MeetingUpdateRequest {
 
     private static final int MAX_NAME_LENGTH = 50;
-    private static final String MISSING_REQUIRED_INPUT = "필수 입력 값이 누락됐습니다.";
 
-    @NotBlank(message = MISSING_REQUIRED_INPUT)
+    @NotBlank(message = ValidationMessages.MISSING_REQUIRED_INPUT)
     @Length(max = MAX_NAME_LENGTH, message = "모임 이름은 " + MAX_NAME_LENGTH + "자를 초과할 수 없습니다.")
     private String name;
