+++ backend/src/main/java/com/woowacourse/moragora/dto/request/user/PasswordRequest.java
@@ -1,5 +1,7 @@
 package com.woowacourse.moragora.dto.request.user;
 
+import com.woowacourse.moragora.constant.Patterns;
+import com.woowacourse.moragora.constant.ValidationMessages;
 import javax.validation.constraints.Pattern;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -12,8 +14,7 @@ public class PasswordRequest {
 
     private String oldPassword;
 
-    @Pattern(regexp = "^(?=.*[0-9])(?=.*[a-zA-Z])(?=.*[!@#&()–[{}]:;',?/*~$^+=<>]).{8,30}$",
-            message = "입력 형식이 올바르지 않습니다.")
+    @Pattern(regexp = Patterns.PASSWORD, message = ValidationMessages.INVALID_FORMAT)
     private String newPassword;
 
     public PasswordRequest(final String oldPassword, final String newPassword) {
