+++ backend/src/main/java/com/woowacourse/moragora/dto/request/user/UserAttendanceRequest.java
@@ -1,5 +1,6 @@
 package com.woowacourse.moragora.dto.request.user;
 
+import com.woowacourse.moragora.constant.ValidationMessages;
 import javax.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -10,9 +11,7 @@ import lombok.ToString;
 @ToString
 public class UserAttendanceRequest {
 
-    private static final String MISSING_REQUIRED_INPUT = "필수 입력 값이 누락됐습니다.";
-
-    @NotNull(message = MISSING_REQUIRED_INPUT)
+    @NotNull(message = ValidationMessages.MISSING_REQUIRED_INPUT)
     private Boolean isPresent;
 
     public UserAttendanceRequest(final Boolean isPresent) {
