+++ backend/src/main/java/com/woowacourse/moragora/dto/request/user/UserDeleteRequest.java
@@ -1,5 +1,6 @@
 package com.woowacourse.moragora.dto.request.user;
 
+import com.woowacourse.moragora.constant.ValidationMessages;
 import javax.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -9,9 +10,8 @@ import lombok.ToString;
 @Getter
 @ToString(exclude = "password")
 public class UserDeleteRequest {
-    private static final String MISSING_REQUIRED_INPUT = "필수 입력 값이 누락됐습니다.";
 
-    @NotBlank(message = MISSING_REQUIRED_INPUT)
+    @NotBlank(message = ValidationMessages.MISSING_REQUIRED_INPUT)
     private String password;
 
     public UserDeleteRequest(final String password) {
