+++ backend/src/main/java/com/woowacourse/moragora/dto/request/user/UserRequest.java
@@ -1,5 +1,7 @@
 package com.woowacourse.moragora.dto.request.user;
 
+import com.woowacourse.moragora.constant.Patterns;
+import com.woowacourse.moragora.constant.ValidationMessages;
 import com.woowacourse.moragora.domain.user.User;
 import com.woowacourse.moragora.domain.user.password.EncodedPassword;
 import javax.validation.constraints.Email;
@@ -13,13 +15,13 @@ import lombok.ToString;
 @ToString(exclude = "password")
 public class UserRequest {
 
-    @Email(regexp = "^[a-zA-Z0-9+-\\_.]+@[a-zA-Z0-9-]+\\.[a-zA-Z0-9-.]+$", message = "입력 형식이 올바르지 않습니다.")
+    @Email(regexp = Patterns.EMAIL, message = ValidationMessages.INVALID_FORMAT)
     private String email;
 
-    @Pattern(regexp = "^(?=.*[0-9])(?=.*[a-zA-Z])(?=.*[!@#&()–[{}]:;',?/*~$^+=<>]).{8,30}$", message = "입력 형식이 올바르지 않습니다.")
+    @Pattern(regexp = Patterns.PASSWORD, message = ValidationMessages.INVALID_FORMAT)
     private String password;
 
-    @Pattern(regexp = "[a-zA-Z0-9가-힣]{1,15}", message = "입력 형식이 올바르지 않습니다.")
+    @Pattern(regexp = Patterns.NICKNAME, message = ValidationMessages.INVALID_FORMAT)
     private String nickname;
 
     public UserRequest(final String email, final String password, final String nickname) {
